separate presentational file, update README and remove opacity style to fix issue with live version
