+++ src/App.css
@@ -66,7 +66,6 @@
 .btn-more-close {
   width: 100%;
   background-color: #4c79f2 !important;
-  opacity: 60%;
 }
 
 .btn-link {
